+++ x/ethbridge/types/codec.go
@@ -8,4 +8,5 @@ import (
 func RegisterCodec(cdc *codec.Codec) {
 	cdc.RegisterConcrete(MsgCreateEthBridgeClaim{}, "ethbridge/MsgCreateEthBridgeClaim", nil)
 	cdc.RegisterConcrete(MsgBurn{}, "ethbridge/MsgBurn", nil)
+	cdc.RegisterConcrete(MsgLock{}, "ethbridge/MsgLock", nil)
 }