+++ x/ethbridge/types/errors.go
@@ -1,6 +1,8 @@
 package types
 
 import (
+	"fmt"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -11,10 +13,12 @@ type CodeType = sdk.CodeType
 const (
 	DefaultCodespace sdk.CodespaceType = ModuleName
 
-	CodeInvalidEthNonce    CodeType = 1
-	CodeInvalidEthAddress  CodeType = 2
-	CodeErrJSONMarshalling CodeType = 3
-	CodeInvalidEthSymbol   CodeType = 4
+	CodeInvalidEthNonce     CodeType = 1
+	CodeInvalidEthAddress   CodeType = 2
+	CodeErrJSONMarshalling  CodeType = 3
+	CodeInvalidEthSymbol    CodeType = 4
+	CodeErrInvalidClaimType CodeType = 5
+	CodeErrInvalidChainID   CodeType = 6
 )
 
 // ErrInvalidEthNonce implements sdk.Error.
@@ -27,6 +31,11 @@ func ErrInvalidEthAddress(codespace sdk.CodespaceType) sdk.Error {
 	return sdk.NewError(codespace, CodeInvalidEthAddress, "invalid ethereum address provided, must be a valid hex-encoded Ethereum address")
 }
 
+// ErrInvalidChainID implements sdk.Error.
+func ErrInvalidChainID(codespace sdk.CodespaceType, chainID string) sdk.Error {
+	return sdk.NewError(codespace, CodeErrInvalidChainID, fmt.Sprintf("invalid ethereum chain id '%s'", chainID))
+}
+
 // ErrJSONMarshalling implements sdk.Error.
 func ErrJSONMarshalling(codespace sdk.CodespaceType) sdk.Error {
 	return sdk.NewError(codespace, CodeErrJSONMarshalling, "error marshalling JSON for this claim")
@@ -35,4 +44,9 @@ func ErrJSONMarshalling(codespace sdk.CodespaceType) sdk.Error {
 // ErrInvalidEthSymbol implements sdk.Error.
 func ErrInvalidEthSymbol(codespace sdk.CodespaceType) sdk.Error {
 	return sdk.NewError(codespace, CodeInvalidEthSymbol, "invalid symbol provided, symbol \"eth\" must have null address set as token contract address")
+}
+
+// ErrInvalidClaimType implements sdk.Error.
+func ErrInvalidClaimType() sdk.Error {
+	return sdk.NewError(DefaultCodespace, CodeErrInvalidClaimType, "invalid claim type provided")
 }