+++ x/ethbridge/types/events.go
@@ -5,12 +5,16 @@ var (
 	EventTypeCreateClaim    = "create_claim"
 	EventTypeProphecyStatus = "prophecy_status"
 	EventTypeBurn           = "burn"
+	EventTypeLock           = "lock"
 
 	AttributeKeyEthereumSender = "ethereum_sender"
 	AttributeKeyCosmosReceiver = "cosmos_receiver"
 	AttributeKeyAmount         = "amount"
 	AttributeKeyStatus         = "status"
+	AttributeKeyClaimType      = "claim_type"
 
+	AttributeKeyEthereumChainID  = "ethereum_chain_id"
+	AttributeKeyTokenContract    = "token_contract"
 	AttributeKeyCosmosSender     = "cosmos_sender"
 	AttributeKeyEthereumReceiver = "ethereum_receiver"