+++ CHANGELOG.md
@@ -52,7 +52,9 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
-- (ethbridge) [\#75](https://github.com/cosmos/peggy/pull/75) Add burn message and functionality for burning tokens and trigerring events.
+- (ethbridge) [\#75](https://github.com/cosmos/peggy/pull/75) Add burn message and functionality for burning tokens and trigerring
+ events.
+- (ethbridge) [\#81](https://github.com/cosmos/peggy/pull/81) Add functionality for locking up cosmos coins and detecting burned coins from ethereum
 
 ### Improvements